requirements.txt
alembic==1.13.2
Flask==3.0.3
Flask_Cors==4.0.1
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
PyJWT==2.8.0
PyJWT==2.8.0
SQLAlchemy==2.0.31
Werkzeug==3.0.3
